Тестування
Повністю протестував бібліотеку і, як виявилося, до бібліотеки потрапив несправна функція. Точніше сказати, справна, але чомусь не хоче працювати... У тестовому файлі, де я пишу алгоритми перед переносом до бібліотеки, при конвертації різних memcell_t отримуються різні int, а у бібліотеці чомусь при конвертації різних memcell_t отримуються однакові int. Чому так? Не знаю. Але я найшов обхідний метод: якщо я буду записувати старші 4 байти до int, а не молодші, як відбувається при стандартній с++ конвертації - отримуються все ж таки різні числа! Все інше у бібліотеці справне. Залишилося лише викотити нове оновлення.

+++ ALGOR.hpp
@@ -78,21 +78,21 @@ namespace ALGOR
 	 * #*****+/^^^/+++++-/+/-+-+                         +-+-/+/-+++++/^^^/+*****# *
 	 * ****+/^^^/+++++-/&/-+-+-+-/&/-+-+-+-/&/-+-+-+-/&/-+-+-+-/&/-+++++/^^^/+**** */
 
-	using sbit8_t = signed char;	   ///< Alias for signed char
-	using sbit16_t = signed short;	   ///< Alias for signed short
-	using sbit32_t = signed long;	   ///< Alias for signed long
-	using sbit64_t = signed long long; ///< Alias for signed long long
+	using sbit8_t = signed char;			///< Alias for signed char
+	using sbit16_t = signed short;			///< Alias for signed short
+	using sbit32_t = signed long;			///< Alias for signed long
+	using sbit64_t = signed long long;		///< Alias for signed long long
 
-	using ubit8_t = unsigned char;		 ///< Alias for unsigned char
-	using ubit16_t = unsigned short;	 ///< Alias for unsigned short
-	using ubit32_t = unsigned long;		 ///< Alias for unsigned long
-	using ubit64_t = unsigned long long; ///< Alias for unsigned long long
+	using ubit8_t = unsigned char;			///< Alias for unsigned char
+	using ubit16_t = unsigned short;		///< Alias for unsigned short
+	using ubit32_t = unsigned long;			///< Alias for unsigned long
+	using ubit64_t = unsigned long long;	///< Alias for unsigned long long
 
-	using fbit32_t = float;	 ///< Alias for float
-	using fbit64_t = double; ///< Alias for double
+	using fbit32_t = float;					///< Alias for float
+	using fbit64_t = double;				///< Alias for double
 
-	using asize_t = unsigned int;		///< Alias for specifying the type "array size"
-	using memcell_t = signed long long; ///< Alias for calculated memory cell
+	using asize_t = unsigned long;			///< Alias for specifying the type "array size"
+	using memcell_t = unsigned long long;	///< Alias for calculated memory cell
 
 	template <typename type_value>
 	void swap(type_value &firstNumber, type_value &secondNumber);
